pyproject.toml
[project]
name = "lyricat-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.31.0",
    "httpx>=0.28.1",
    "lyricsgenius>=3.7.2",
    "pydantic>=2.11.7",
    "uvicorn>=0.35.0",
]
